requirements.txt
pytest==7.4.2
selenium==4.37.0
pyyaml
webdriver-manager
allure-pytest==2.12.0
openpyxl
mss